Avoid implicit floating point conversions [v2]
These were found with clang -Wdouble-promotion and show places where
floating point values were being implicitly converted between
representations. These conversions can result in unexpected use of
double precision arithmetic. Those which are intentional all have an
explicit cast added.


v2:
	Use clang built-in classification functions.
	Declare constants as 'float' instead of double.

+++ newlib/libm/common/cosf.c
@@ -41,7 +41,7 @@
 float
 cosf (float y)
 {
-  double x = y;
+  double x = (double) y;
   double s;
   int n;
   const sincos_t *p = &__sincosf_table[0];
