+++ newlib/libm/common/exp2.c
@@ -112,9 +112,9 @@ exp2 (double x)
 	return WANT_ROUNDING ? 1.0 + x : 1.0;
       if (abstop >= top12 (1024.0))
 	{
-	  if (asuint64 (x) == asuint64 (-INFINITY))
+	  if (asuint64 (x) == asuint64 ((double) -INFINITY))
 	    return 0.0;
-	  if (abstop >= top12 (INFINITY))
+	  if (abstop >= top12 ((double) INFINITY))
 	    return 1.0 + x;
 	  if (!(asuint64 (x) >> 63))
 	    return __math_oflow (0);
